# DOC/api/openapi.yaml
openapi: 3.1.0
info:
  title: netdisk REST API
  version: 1.0.0
  description: |
    网盘系统的**唯一接口契约**(11 章:servergen 由它生成 web 与桌面端的模型)。

    三条纪律:

    1. **本文件是唯一来源**。任何接口改动必须改这里,然后 `pnpm gen:api` 重新生成
       客户端类型;CI 用 `pnpm check:api` 做**双向 diff 门禁** —— 契约改了没重新生成、
       或生成的类型被手改过,都会在这里被拦下。
    2. **手写类型被 lint 拦**(`pnpm lint:contract`):`packages/api` 之外不得出现
       `fetch(`,也不得手写与生成模型同名的对象形状 —— 否则"契约单源"名存实亡。
    3. **错误体统一**为 `Error`(与 internal/apierr 的 JSON 一一对应)。
       客户端按 `code` 分流,绝不解析 `message` 文案。

    覆盖范围:一期已实现且前端/桌面端会调用的接口。WebDAV 是 HTTP 协议入口,
    不进本契约(它的语义由 RFC 4918 定义);`/metrics`、`/healthz` 属运维端点,同样不入。
servers:
  - url: /
    description: 与前端同源(Go 单端口;Nginx 只做 TLS 与分流)
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 登录(账密)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: 刷新令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/files:
    get:
      tags: [files]
      summary: 列目录(keyset 分页)
      description: |
        参数名以本文档与架构文档为准:`space` / `parent`。
        服务端同时接受历史拼法 `space_id` / `parent_id`(同义别名),
        但**只声明一个规范名** —— 声明两个同义参数会让生成的客户端类型出现
        两个字段,调用方无从选择。
      parameters:
        - { name: space, in: query, schema: { type: string } }
        - { name: parent, in: query, schema: { type: string } }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
        - name: after
          in: query
          description: 上一页最后一个条目的 lower(name)
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileListResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "410": { $ref: "#/components/responses/SpaceRevoked" }
  /api/v1/files/dirs:
    post:
      tags: [files]
      summary: 新建目录(Bearer 令牌可用的 REST 入口)
      description: |
        目录创建此前**只有 WebDAV MKCOL** 一个入口(见 ADR-6:WebDAV 走 Basic
        账密换票)。桌面客户端按设计只持有 Bearer 令牌、没有账密,于是它
        **建不了目录** —— 连带"任何要放进子目录的文件"都传不上去
        (实测报错:`WebDAV 认证失败: 请提供 Basic 凭据`)。本路径补上这个缺口。

        `space_id` / `parent_id` 缺省时分别落到**调用者的个人空间**与**该空间
        根目录**(与 `GET /api/v1/files` 的两个参数同义,含历史别名拼法)。

        语义是**严格新建**:同目录下已有同名条目(不分大小写)→ 409
        `name_conflict`,**不**顺手返回已有条目 —— 想"复用已存在目录"的调用方
        自己按名字查一次;把两种意图塞进同一个 201 会让客户端无法区分
        "我刚建了它"与"它本来就存在",而这正是同步逻辑需要的区别。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                space_id: { type: string, description: 缺省 = 调用者个人空间 }
                parent_id: { type: string, description: 缺省 = 空间根目录 }
                name: { type: string }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntryView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { $ref: "#/components/responses/SpaceRevoked" }
  /api/v1/files/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [files]
      summary: 文件详情
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntryView" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [files]
      summary: 改名(带 base_version 乐观锁)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                base_version: { type: integer, format: int64, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntryView" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [files]
      summary: 硬删(无回收站;目录删整棵子树)
      responses:
        "200":
          description: 同步完成(子树行数 ≤ dir_op_sync_max_rows)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
        "202":
          description: |
            超阈值已转异步删除(6.11):行**此刻还在**,响应里只有 task_id。
            轮询 `GET /api/v1/tasks/{id}`;入队后该子树上的读写一律 409
            `dir_op_in_progress`,直到任务结束。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/files/{id}/content:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    head:
      tags: [files]
      summary: 取内容元信息(与 GET 同头:ETag/Content-Length/Content-Type,不回 body)
      responses:
        "200": { description: 头信息 }
        "304": { description: ETag 未变 }
        "404": { $ref: "#/components/responses/NotFound" }
    get:
      tags: [files]
      summary: 下载(流式;支持 Range/If-Range/条件请求)
      responses:
        "200": { description: 完整内容 }
        "206":
          description: 部分内容
          headers:
            Content-Range:
              schema: { type: string }
        "304": { description: ETag 未变 }
        "404": { $ref: "#/components/responses/NotFound" }
        "416": { description: Range 越界 }
  /api/v1/files/{id}/move:
    post:
      tags: [files]
      summary: 移动(可同时改名;不跨空间)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parent_id]
              properties:
                parent_id: { type: string }
                name: { type: string }
                base_version: { type: integer, format: int64, default: 0 }
      responses:
        "200":
          description: 同步完成(子树行数 ≤ dir_op_sync_max_rows)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MoveResult" }
        "202":
          description: |
            超阈值已转异步任务(6.11):响应里只有 task_id,条目还在原位置。
            客户端轮询 `GET /api/v1/tasks/{id}`,任务完成后再刷新列表 ——
            **不要**在收到 202 时本地做乐观移动(任务可能失败)。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MoveResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/tasks/{id}:
    get:
      tags: [files]
      summary: 目录级异步任务状态(6.11,供轮询)
      description: |
        只返回**自己的**任务(管理员例外);非本人一律 404 而不是 403 ——
        403 等于确认"这个任务 id 存在",而任务里有空间/文件 id 与文件名。
        未结束时**不回** `rows_affected/freed_bytes/released_objects`
        (回 0 会被客户端当成"这次操作什么都没删")。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/files/{id}/subtree-stats:
    get:
      tags: [files]
      summary: 子树影响面预检(删除确认与容量预估共用)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubtreeStats" }
  /api/v1/files/{id}/copy:
    post:
      tags: [files]
      summary: 复制文件或目录(逻辑复制 + 引用计数,不搬内容)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                target_parent_id: { type: string }
                name: { type: string, description: 空则自动生成"原名 (副本)" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CopyResult" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/files/{id}/share-to-space:
    post:
      tags: [files]
      summary: 共享到另一个空间(只新建元数据行 + 引用 +1)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_space_id]
              properties:
                target_space_id: { type: string }
                target_parent_id: { type: string }
                name: { type: string }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntryView" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/files/{id}/lock:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [files]
      summary: 查询编辑锁(不含持有者身份以外的信息;无锁时 404)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  lock: { $ref: "#/components/schemas/LockView" }
                  ttl_seconds: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [files]
      summary: 获取/续期编辑锁(30min;过期可抢占)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  lock: { $ref: "#/components/schemas/LockView" }
                  ttl_seconds: { type: integer, format: int64 }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [files]
      summary: 释放编辑锁(幂等)
      responses:
        "204": { description: 已释放 }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/upload/create:
    post:
      tags: [upload]
      summary: 创建上传任务(预留额度;预检命中时下发秒传挑战)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadCreateRequest" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadCreateResult" }
        "409": { $ref: "#/components/responses/Conflict" }
        "507": { $ref: "#/components/responses/QuotaExceeded" }
  /api/v1/upload/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    delete:
      tags: [upload]
      summary: 取消上传(即时释放预留)
      responses:
        "204": { description: 已取消 }
    head:
      tags: [upload]
      summary: 取当前偏移(断点续传)
      responses:
        "200":
          description: OK
          headers:
            Upload-Offset: { schema: { type: integer } }
            Upload-Length: { schema: { type: integer } }
  /api/v1/upload/{id}/finish:
    post:
      tags: [upload]
      summary: 秒传定稿(免重传;证明未过 → 403 fast_upload_denied)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nonce, sample_sha256]
              properties:
                nonce: { type: string }
                sample_sha256:
                  type: array
                  items: { type: string }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadSimpleResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/changes:
    get:
      tags: [sync]
      summary: 增量变更流(超窗 → 409 cursor_expired → 全量清单)
      parameters:
        - { name: space, in: query, required: true, schema: { type: string } }
        - { name: since, in: query, schema: { type: integer, format: int64, default: 0 } }
        - name: limit
          in: query
          schema: { type: integer, maximum: 5000, default: 1000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChangesPage" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { $ref: "#/components/responses/SpaceRevoked" }
  /api/v1/changes/head:
    get:
      tags: [sync]
      summary: 当前最大 change_seq(首次同步的增量起点)
      parameters:
        - { name: space, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  space_id: { type: string }
                  head_seq: { type: integer, format: int64 }
  /api/v1/sync/cursors:
    post:
      tags: [sync]
      summary: 上报游标(单维取 max;全局水位取 min)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, space_id, last_seq]
              properties:
                client_id: { type: string }
                space_id: { type: string }
                last_seq: { type: integer, format: int64 }
      responses:
        "200": { description: OK }
...
  /api/v1/admin/departments/{id}:
    delete:
      tags: [admin]
      summary: 删除部门
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: 未认证/令牌无效(need_password=true 表示分享需要密码)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login: { type: string }
        password: { type: string }
        audience:
          type: string
          description: 分端受众(R-14):web / h5 / desktop
          enum: [web, h5, desktop]
...
    LoginResult:
      type: object
      required: [token, user]
      properties:
        token: { $ref: "#/components/schemas/TokenPair" }
        user: { $ref: "#/components/schemas/UserBrief" }
        personal_space: { $ref: "#/components/schemas/SpaceBrief" }
...
security:
  - bearerAuth: []